Refactor utils from cluster
