+++ framework/wazuh/core/server/unix_server/tests/test_config.py
@@ -34,7 +34,7 @@ mock_config_data = {
 
 @pytest.fixture
 def patch_load():
-    """Patch the load method in CentralizedConfig"""
+    """Patch the load method in CentralizedConfig."""
     with patch.object(CentralizedConfig, 'load', return_value=None):
         with patch.object(ValidateFilePathMixin, '_validate_file_path', return_value=None):
             CentralizedConfig._config = Config(**mock_config_data)
