+++ framework/wazuh/core/server/unix_server/tests/test_server.py
